requirements.txt
Flask==3.0.2
python-dotenv
openai
pymongo
flask-mail
requests
web3==5.31.3
binance
python-binance
paypalrestsdk
hyperwallet-sdk
flask-wtf==1.2.1
wtforms==3.1.2
celery
gunicorn
pandas
openpyxl
email-validator
bs4
bcrypt
Flask-SQLAlchemy
psycopg2-binary
apscheduler
Flask-HTTPAuth
qrcode==7.4.2
pydub
pytz
azure-cognitiveservices-speech
markdown
werkzeug
#flask-socketio  # Uncomment if using Flask-SocketIO
beautifulsoup4
colorama
pypng
typing-extensions
зфердши